sniper.py
"""
EVM Mempool Sniper Bot
======================
A professional liquidity sniping bot for Uniswap V4 and compatible DEXs.

Features:
- Async WebSocket mempool monitoring
- PairCreated event detection
- Honeypot safety checks
- Flashbots Protect integration
- EIP-1559 gas optimization
- Multi-chain support (ETH, Base, Arbitrum)

Author: Inan
Version: 1.0.0
License: MIT

⚠️ DISCLAIMER: This is for educational/portfolio purposes only.
   Trading crypto carries significant financial risk.
"""

import asyncio
import os
import sys
from datetime import datetime
from typing import Optional

from dotenv import load_dotenv

# Fix Windows console encoding for emoji support
if sys.platform == "win32":
    sys.stdout.reconfigure(encoding='utf-8')

# Load environment variables
load_dotenv()


# =============================================================================
# Configuration
# =============================================================================

class Config:
    """Bot configuration loaded from environment variables."""
    
    # RPC Endpoints
    ETH_WSS: str = os.getenv("ETH_WSS", "")
    BASE_WSS: str = os.getenv("BASE_WSS", "")
    ARB_WSS: str = os.getenv("ARB_WSS", "")
    
    # Wallet
    PRIVATE_KEY: str = os.getenv("PRIVATE_KEY", "")
    
    # Flashbots
    FLASHBOTS_RELAY: str = os.getenv(
        "FLASHBOTS_RELAY", 
        "https://relay.flashbots.net"
    )
    
    # Contract Addresses
    UNISWAP_V4_POOL_MANAGER: str = "0x000000000004444c5dc75cB358380D2e3dE08A90"
    UNISWAP_V2_FACTORY: str = "0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f"
    WETH: str = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2"
    
    # Trading Parameters
    SNIPE_AMOUNT_ETH: float = 0.5
    SLIPPAGE_PERCENT: float = 5.0
    GAS_LIMIT: int = 300000
    PRIORITY_FEE_GWEI: float = 2.5
    
    # Safety
    HONEYPOT_API: str = "https://api.honeypot.is/v2/IsHoneypot"
    MAX_BUY_TAX: float = 10.0
    MAX_SELL_TAX: float = 10.0


# =============================================================================
# ABIs (Minimal for demo)
# =============================================================================

PAIR_CREATED_ABI = {
    "anonymous": False,
    "inputs": [
        {"indexed": True, "name": "token0", "type": "address"},
        {"indexed": True, "name": "token1", "type": "address"},
        {"indexed": False, "name": "pair", "type": "address"},
        {"indexed": False, "name": "allPairsLength", "type": "uint256"},
    ],
    "name": "PairCreated",
    "type": "event",
}


# =============================================================================
# Utility Functions
# =============================================================================

def log(emoji: str, message: str) -> None:
    """Print a formatted log message with timestamp."""
    timestamp = datetime.now().strftime("%H:%M:%S")
    print(f"[{timestamp}] {emoji} {message}")


def log_header() -> None:
    """Print the bot header."""
    print("\n" + "=" * 60)
    print("  EVM MEMPOOL SNIPER BOT v1.0.0")
    print("  Uniswap V4 | Flashbots Protect | Multi-Chain")
    print("=" * 60 + "\n")


# =============================================================================
# Honeypot Detection
# =============================================================================

async def check_honeypot(token_address: str, chain: str = "eth") -> dict:
    """
    Check if a token is a honeypot using honeypot.is API.
    
    Args:
        token_address: The token contract address
        chain: Chain identifier (eth, bsc, base)
    
    Returns:
        dict with honeypot status and tax info
    """
    try:
        import aiohttp
        
        url = f"{Config.HONEYPOT_API}?address={token_address}&chainId=1"
        
        async with aiohttp.ClientSession() as session:
            async with session.get(url, timeout=5) as response:
                if response.status == 200:
                    data = await response.json()
                    return {
                        "is_honeypot": data.get("honeypotResult", {}).get("isHoneypot", False),
                        "buy_tax": data.get("simulationResult", {}).get("buyTax", 0),
                        "sell_tax": data.get("simulationResult", {}).get("sellTax", 0),
                        "is_safe": not data.get("honeypotResult", {}).get("isHoneypot", True),
                    }
    except Exception as e:
        log("⚠️", f"Honeypot check failed: {e}")
    
    return {"is_honeypot": True, "is_safe": False, "buy_tax": 100, "sell_tax": 100}


# =============================================================================
# Gas Optimization
# =============================================================================

def calculate_gas_params(base_fee_gwei: float) -> dict:
    """
    Calculate EIP-1559 gas parameters.
    
    Args:
        base_fee_gwei: Current base fee in Gwei
    
    Returns:
        dict with maxFeePerGas and maxPriorityFeePerGas in Wei
    """
    priority_fee_wei = int(Config.PRIORITY_FEE_GWEI * 1e9)
    max_fee_wei = int((base_fee_gwei + Config.PRIORITY_FEE_GWEI * 2) * 1e9)
    
    return {
        "maxFeePerGas": max_fee_wei,
        "maxPriorityFeePerGas": priority_fee_wei,
        "gas": Config.GAS_LIMIT,
    }


# =============================================================================
# Flashbots Integration (Conceptual)
# =============================================================================

async def send_flashbots_bundle(signed_tx: str) -> Optional[str]:
    """
    Send a transaction via Flashbots Protect for frontrun protection.
    
    Args:
        signed_tx: The signed transaction hex
    
    Returns:
        Transaction hash if successful, None otherwise
    
    Note:
        This is a conceptual implementation for portfolio demonstration.
        Production use requires proper Flashbots SDK integration.
    """
    try:
        import aiohttp
        
        payload = {
            "jsonrpc": "2.0",
            "id": 1,
            "method": "eth_sendPrivateTransaction",
            "params": [{"tx": signed_tx}],
        }
        
        async with aiohttp.ClientSession() as session:
            async with session.post(
                Config.FLASHBOTS_RELAY,
                json=payload,
                headers={"Content-Type": "application/json"},
            ) as response:
                if response.status == 200:
                    data = await response.json()
                    return data.get("result")
    except Exception as e:
        log("❌", f"Flashbots submission failed: {e}")
    
    return None


# =============================================================================
# Transaction Execution (Conceptual)
# =============================================================================

async def execute_snipe(token_address: str, amount_eth: float) -> bool:
    """
    Execute a snipe transaction.
    
    Args:
        token_address: Target token to buy
        amount_eth: Amount of ETH to spend
    
    Returns:
        True if successful
    
    Note:
        This is a simulation for portfolio demonstration.
    """
    log("💰", f"Preparing snipe: {amount_eth} ETH → {token_address[:10]}...")
    
    # Simulate gas calculation
    base_fee = 42.0  # Would come from w3.eth.get_block('latest')
    gas_params = calculate_gas_params(base_fee)
    
    log("⛽", f"Gas: {base_fee:.1f} Gwei | Priority: {Config.PRIORITY_FEE_GWEI} Gwei")
    log("🚀", "Submitting via Flashbots Protect...")
    
    # Simulate transaction
    await asyncio.sleep(0.5)
    
    fake_hash = "0x3b8a7c9d2e1f4a5b6c7d8e9f0a1b2c3d4e5f6a7b"
    log("✅", f"Private TX Sent: {fake_hash[:10]}...{fake_hash[-4:]}")
    log("🛡️", "MEV Protection: ACTIVE")
    
    return True


# =============================================================================
# Event Handlers
# =============================================================================

async def handle_pair_created(
    token0: str, 
    token1: str, 
    pair: str,
) -> None:
    """
    Handle a new PairCreated event.
    
    Args:
        token0: First token address
        token1: Second token address  
        pair: Pair contract address
    """
    # Determine which token is the new one (not WETH)
    weth_lower = Config.WETH.lower()
    if token0.lower() == weth_lower:
        target_token = token1
    elif token1.lower() == weth_lower:
        target_token = token0
    else:
        log("⏭️", "Skipping non-WETH pair")
        return
    
    log("⚡", f"NEW PAIR: WETH/{target_token[:10]}...")
    
    # Honeypot check
    log("🔒", "Running honeypot check...")
    hp_result = await check_honeypot(target_token)
    
    if hp_result["is_safe"]:
        log("🔒", f"Honeypot Check: ✅ SAFE (Buy: {hp_result['buy_tax']:.1f}%, Sell: {hp_result['sell_tax']:.1f}%)")
        
        # Check tax limits
        if hp_result["buy_tax"] <= Config.MAX_BUY_TAX and hp_result["sell_tax"] <= Config.MAX_SELL_TAX:
            await execute_snipe(target_token, Config.SNIPE_AMOUNT_ETH)
        else:
            log("⚠️", "Tax too high, skipping")
    else:
        log("🔒", "Honeypot Check: ❌ UNSAFE - Skipping")


# =============================================================================
# Mempool Monitor (Simulation)
# =============================================================================

async def simulate_mempool_monitor() -> None:
    """
    Simulate mempool monitoring for portfolio demonstration.
    
    In production, this would use:
    - web3.eth.subscribe('newPendingTransactions')
    - web3.eth.subscribe('logs', {address: FACTORY})
    """
    log("📡", "Subscribed: newPendingTransactions")
    log("🔍", "Monitoring Uniswap V4 PoolManager...")
    log("🔍", "Monitoring Uniswap V2 Factory...")
    
    # Simulate scanning blocks
    block_num = 19842103
    for _ in range(3):
        await asyncio.sleep(1)
        log("🔍", f"Scanning Block #{block_num}...")
        block_num += 1
    
    # Simulate finding a new pair
    await asyncio.sleep(0.5)
    
    # Trigger demo pair event
    demo_token = "0x6B175474E89094C44Da98b954EedeAC495271d0F"
    await handle_pair_created(
        token0=Config.WETH,
        token1=demo_token,
        pair="0x1234567890abcdef1234567890abcdef12345678",
    )


# =============================================================================
# Main Entry Point
# =============================================================================

async def main() -> None:
    """Main entry point for the sniper bot."""
    log_header()
    
    # Validate configuration
    if not Config.ETH_WSS and not os.getenv("DEMO_MODE"):
        log("ℹ️", "Running in DEMO mode (no RPC configured)")
    
    # Display connection status
    log("🔗", "Connected: Ethereum Mainnet [Block #19842103]")
    
    # Run the monitor
    await simulate_mempool_monitor()
    
    print("\n" + "=" * 60)
    log("✨", "Demo complete! Check README.md for production setup.")
    print("=" * 60 + "\n")


if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("\n[!] Bot stopped by user")
        sys.exit(0)
